+++ packages/core/__tests__/Hub-test.ts
@@ -88,6 +88,25 @@ describe('Hub', () => {
 		);
 	});
 
+	test('Protected channel - ui', () => {
+		const listener = jest.fn(() => {});
+		const loggerSpy = jest.spyOn(Logger.prototype, '_log');
+
+		Hub.listen('ui', listener);
+
+		Hub.dispatch('ui', {
+			event: 'auth:signOut:finished',
+			data: 'the user has been signed out',
+			message: 'User has been signed out',
+		});
+
+		expect(listener).toHaveBeenCalled();
+		expect(loggerSpy).toHaveBeenCalledWith(
+			'WARN',
+			'WARNING: ui is protected and dispatching on it can have unintended consequences'
+		);
+	});
+
 	test('Regex Listener', () => {
 		const listener = jest.fn(() => {});