+++ packages/analytics/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [5.1.12](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/analytics@5.1.11...@aws-amplify/analytics@5.1.12) (2022-02-03)
+
+**Note:** Version bump only for package @aws-amplify/analytics
+
+
+
+
+
+## [5.1.11](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/analytics@5.1.10...@aws-amplify/analytics@5.1.11) (2022-01-27)
+
+**Note:** Version bump only for package @aws-amplify/analytics
+
+
+
+
+
+## [5.1.10](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/analytics@5.1.9...@aws-amplify/analytics@5.1.10) (2022-01-07)
+
+**Note:** Version bump only for package @aws-amplify/analytics
+
+
+
+
+
 ## [5.1.9](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/analytics@5.1.8...@aws-amplify/analytics@5.1.9) (2021-12-16)
 
 **Note:** Version bump only for package @aws-amplify/analytics